Smoke Flowers emit Smoke now
The Smoke Flowers i created are able to emit smoke now upon being interacted with. The smoke doesn't appear the way I wanted it to however since Niagara Systems work using a 2D space so they follow the camera which makes it hard to appear as if the smoke is covering the area and will be hard to make it seem like its blocking the enemy's view like this. I need someone to test them and give me feedback on them