Added tutorial spaces into the map
I added the tutorial spaces into the map in the spots i believe they may be of use. Some of them however will use the wrong text, this is intentional as they're meant just be a reminder of where they tutorial volumes are as well as to remind the people making the tutorials to add more tutorials.