Removed Memory Puzzle Room Enemies
Removed the enemies from the memory puzzle room (room with trees). They were there in every branch i made and in the build despite me being told that they were supposed to be removed so i removed them. I'm not sure if this will work for everyone but hopefully it does.